backend/requirements.txt
# Backend dependencies for Hymns OBS
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
aiosqlite==0.20.0

# YAML parsing
pyyaml==6.0.1

# WebSockets for real-time communication
websockets==12.0

# CORS
python-multipart==0.0.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.27.0

# Development
python-dotenv==1.0.1
